+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Minor bugfixes for existing SICD consistency `check_*` methods
+
+### Added
+- New `check_*`methods for SICD consistency
+- More complete unit tests for SICD consistency
+
 
 ## [1.0.1] - 2025-09-09