+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `ElementWrapper.get` now supports the `default` parameter
+
 ### Changed
 - `ElementWrapper.__contains__` now raises `KeyError` for impossible keys