+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `sarkit.cphd.Reader` support for reading a subset of vectors
 
+### Changed
+- `ElementWrapper.__contains__` now raises `KeyError` for impossible keys
+
 
 ## [1.0.1] - 2025-09-09