+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `sarkit.cphd.Reader` support for reading a subset of vectors
+
 ### Changed
 - `ElementWrapper.__contains__` now raises `KeyError` for impossible keys
+
 
 ## [1.0.1] - 2025-09-09