+++ CHANGELOG.md
@@ -9,9 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `sarkit.cphd.Reader` support for reading a subset of vectors
+- New `check_*`methods for SICD consistency
+- More complete unit tests for SICD consistency
 
 ### Changed
 - `ElementWrapper.__contains__` now raises `KeyError` for impossible keys
+
+### Fixed
+- Minor bugfixes for existing SICD consistency `check_*` methods
 
 
 ## [1.0.1] - 2025-09-09