+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [1.1.0] - 2025-09-29
+
 ### Added
 - `sarkit.cphd.Reader` support for reading a subset of vectors
 - New `check_*`methods for SICD consistency